app/requirements.txt
markupsafe==2.0.1
flask==1.1.4
flask-migrate
python-dotenv==0.19.2
mariadb==1.0.9
sqlalchemy==1.4.29